Add tests and ITs
